greed.md: Deleted.
I've thought about whether to rewrite this (and did have a draft sitting
around), or to just disavow it altogether and delete it. I decided that
this blog isn't really meant to be about politics; revising this to more
concisely/accurately describe my beliefs will probably be more hassle
than it's worth. Away this article goes.