Adopt selenized as website's color scheme.
There are some issues with the previous dark color scheme in particular.
My thought was to go above and beyond and have link colors maintain a
3:1 foreground-to-link ratio and 4.5:1 foreground-to-background ratio.
Only the latter is actually required since I underline links.
Unfortunately, in maintaining that strict requirement, it had forced
the foreground and background colors to be far apart, which made
reading the main body more difficult than it had to be due to things
like halation (I personally deal with halation so in that sense, I can
test for it).

I'm also not a "color person." I know a little about it, not a lot, and
I feel it's better to draw on the knowledge of people more qualified
than I am. There was a lot of time and effort spent on solarized, and a
lot of time and effort spent on selenized. I simply am not in a position
to invest that much of either into a unique color scheme. This one works
well enough for my purposes.

This finally patches stagit colors as well, which has been bothering me
for quite a while. I don't know that I'll stick with stagit, but we'll
see.

Oh. Due to the caching HTTP headers I've set in the backend, it's
possible that some people will still see the old stagit colors. Not
immediately sure how to fix this yet, I'll have to look at how stagit
handles CSS files to see if I can cache-bust or not.