chore(fonts): migrate to Source Sans 3
* Replace Assistant with Source Sans 3 as the main font
* Add italic variants for both Source Sans 3 and Source Code Pro
* Update all font files and CSS references accordingly

Assistant has some small bugs that are quite bothersome. One being that
the j is slightly more bold than the other characters due to an
overlapping vector. Also, there is no italics support which means that
the browser will synthesize italics, and I don't necessarily like that.
But, fortunately Source Sans 3 is a very similar font, basically the
same font, and it solves these issues.

Also, the fonts should be smaller now. I used the uharfbuzz binary, used
OTF files instead of TTF files to make the WOFF2 fonts, and I also
cut out frac/dnom/numr support. They're all down to around 20K file size
(which is decent for a custom font).

In the future I could potentially use a fancier technique with
javascript to do two stage loading, but this is good enough for now.
Also would want to design that carefully so that users that disable
javascript will still get the full experience. Would try to keep it
minimal if I did so, and I know some readers in my audience probably
wouldn't like the use of _any_ javascript so I have to keep that in mind
as well.

The big motivations for moving to custom fonts rather than just using
system fonts are these:

- consistency between devices. Every device gets the same fonts.
- more professional looking (which matters if potential employers look
  at the site, and the reality is that I have to care about that).
- consistency between fonts. It looks janky if the monospace font and
  the sans-serif font are completely mismatched. There's a sense of
  unity with these two fonts that is really nice.
- I know that the fonts are decent choices from a technical standpoint.

The drawbacks are:

- users will have to block the remote fonts using ublock origin or
  something similar if they want to use their own, I think. Wish there
  was a way I could have both, this part is unfortunate. :(

- users on slower connections may have a tougher time. I've done what I
  can right now to mitigate this, but still a factor. This is also
  unfortunate. I know I could do more and it's something I can look into
  when I have more time. I didn't put anything in the critical rendering
  path, so users on slow connections should be okay in this regard, and
  they can block the remote fonts as well. Hate to have that as a
  solution but I don't know what else to do there.

- users on limited data plans would most likely have a difficult time as
  well. But at that point, I shouldn't be using images much either.

- flash of unstyled text (FOUT). Actually another motivation to work
  more on the above point. The two stage loading thing I read about
  would help. Would result in slightly more data getting downloaded
  overall, however. This is most likely a worse situation over the onion
  service, too.

Bit of a long commit message, but figured I would write a bit about my
motivations in case anyone actually reads these. I do care about the
users that visit my site and want them to have a good experience, I just
also have to balance that with my need to get hired.