feat(fonts): optimize font loading with subset files
- Add subset WOFF2 font files for Source Sans 3 and Source Code Pro
- Remove unicode-range restrictions to two files
- Update CSS file references across all HTML pages

Point here is to make the files as small as possible while still
covering all of the major code points. The subset files aren't even
variable fonts - they only have a weight of 400 available. So faux bold
and italic will show. But this is ultimately not a bad thing because
each file is around 5-6 kilobytes due to removing variable font support
and being aggressively subset - around 1/4 the size of the full files,
which will be loaded after.