fix(fonts): add () to sans font, regen styles
While here, add unicode-range back in for Source Sans 3 and Source Code
Pro. I'm hoping that by adding these back in, it will force the page to
wait until it can determine which characters are on it, at which point
the more aggressively subset fonts will (hopefully) have already loaded.

I think I'm going to have to use a small amount of javascript at some
point to ensure that the fonts have been loaded. But then I guess I'll
be changing the site slightly so that it loads the full latin set if
javascript is disabled, that way those users still have proper fonts.
They might see a flash of unstyled text instead of a flash of faux text,
but that's the trade off for disabling javascript in that situation
unfortunately.