fix(fonts): load all variants for source code/sans
Made a bit of a silly mistake. Problem with the last approach was,
source code pro and italics variants may never actually be loaded
properly by the JS. So in the case of our monospace font, it may have a
FOUT anyway. That is unfortunate, because that's the entire purpose of
the JS - to handle FOUT by either eliminating or making it as brief as
possible, replacing it with FOFT instead.

The intention behind the last approach was to optimize the font loading
so that the number of requests and needed bandwidth were minimized. But
handling FOUT is more important in my opinion.

For posterity, let's outline the bandwidth differences between before
and after custom fonts were implemented on this site:

- About ~11KB for the preloaded fonts
- ~85KB for the four latin versions I'm loading here

So ~96KB total there.

I support many other character sets - extended latin, cyrillic,
vietnamese, greek, and other. But these are all loaded as needed with
unicode-range since my writing itself probably wouldn't trigger them -
it's mostly in case some translation software decides to rewrite things
for a user in a different country. So I wouldn't count these - they are
probably never loaded for most users. I have them "just in case". The
web version of Google Translate, for instance, does not even use them.

Then I suppose I should also count the stylesheet bloat caused by
including all these font faces. Uncompressed, that would be about ~20KB.
But it compresses down to ~3.5KB, and my build system for the server
will automatically do that. So I would count that as ~100KB total. JS
file size is negligible - about a KB.

I am still on the fence whether I should be loading the custom fonts for
the non-JavaScript users, knowing full well they will get a FOUT; or, if
I should do sans-serif or a web-safe font instead, knowing full well it
could look ugly.

The important thing to note here is, the added critical amount to load
is much less than the total amount.

- ~11KB for preload fonts.
- ~3.5KB for gzipped CSS.

About 15KB. Not bad. Worse than 0 of course, but not terrible
considering the gains in aesthetics, professional appearance, and visual
consistency.

Also, I'm not inlining anything, so those that block remote fonts or use
CLI web browsers shouldn't see much difference in data transferred.