style(fonts): remote fonts blocked -> sans-serif
Not sure how I feel about falling back to Arial and Courier New. I only
did this when I was planning to set an adjusted fallback font that
matched the proportions of the source sans/code fonts. But since that
isn't the case now, I don't see a purpose in overriding sans-serif.

Some frameworks use system font stacks, but I don't want to prematurely
switch to what everyone else is using. If I find a good reason to do it,
I'll switch. If not, I'll just stick with these two fonts.