feat(colors): fully APCA compliant light theme
Took a lot of tweaking but I finally got it. Even code blocks and links
are APCA compliant now!

APCA is _much_ more strict compared to WCAG AAA. On the bright side, the
theme should be much more accessible to colorblind people and those with
deficiencies, as well as much more readable in broad daylight.

May adjust heading-fg-color, not 100% happy with it yet but want to push
this all the same.