style: remove size-adjust and increase font size
- Remove size-adjust: 106.5% from all @font-face declarations
- Increase --base-font-size from clamp(1rem, 0.913rem + 0.435vw, 1.25rem)
  to clamp(1.065rem, 0.973rem + 0.463vw, 1.331rem)
- Update CSS file hash in all HTML files to reflect changes

Ended up finding out that line height is too low with this font. So I'm
going to have to fix this. This is a quick fix, but maybe not the most
correct fix in the long run (what if I incorporate a heading font? Then
this little hack won't work anymore). But it's better to fix it now the
imperfect way than have it broken until a perfect fix.