style: fix line length - max 80 characters
So my line length was way out of wack back when I was using 66ch as the
measurement. I changed this to 38rem yesterday, knowing that I'd
probably come back and bump it down more. This is that change. I did
some testing with AI generated javascript and `wc` and found that 35rem
is the point at which lines are never longer than 81 characters
including a trailing space. Which is what I want for WCAG AAA visual
presentation (SC 1.4.8).