style: use a px based system for margins, not rem
Also adjusted the thresholds using a CSS outline as a guide. On the very
pixel that `margin: auto;` starts to kick in (629px), we set the
breakpoint to start decreasing the padding. The very moment that the
background starts to appear on the other side of the 24px outline
(686px), that is the point at which our padding should be 0 since the
margins.

I like this new way of doing it much better. The problem with rem is,
when using a fluid typography system the size of the type influences the
borders. Which in retrospect should have been obvious last night, but I
was tired and falling back on "don't use px for responsive design."